requirements.txt
tensorflow
Pillow
PyUserInput
selenium